Breaking: Make it more clear why a compression method is recommended
Just specifying that it "Should be served compressed with Zopfli"
when the user is also informed that the same resource "Should be
served compressed with Brotli over HTTPS" is confusing.

In order to make it more clear, this commit changes the error messages
so that they provide a bit more information, thus, it changes something
such as "Should be served compressed with Zopfli" to "Should be served
compressed with Zopfli when gzip compression is requested".

- - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - -

Fix #913
Fix #914

Close #915