Chore: Automatically merge Greenkeeper's commits
The `greenkeeper-merge.js` script is executed as a job when all
the tests passed. `Greenkeeper` does a commit per each package
in the monorepo that shares the same dependency.

The script:

* Squashes all of those commits (takes all the commits from the
  end with the same message until one is different).
* Rebases the changes into `master` and pushes.

**NOTE**: The `install` step of `travis.yml` performs a `yarn`.
If the version is different, the generated `yarn.lock` file will
be too. This script discards any changes in the `yarn.lock` that
originate in that step (`Greenkeeper` takes care of updating it).

- - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - -

Fix #938
Close #1313

Co-authored-by: Cătălin Mariș <alrraa@gmail.com>