Fix: Disable hints that are currently noisy when run locally
Disable `hint-apple-touch-icons` because it is not aware if the
project is a PWA or not (needs to check for a manifest).

Disable `webpack-config/module-esnext-typescript` because it
requires loading both the webpack and typescript configs at the
same time to work (which doesn't currently happen in VS Code).

These can be re-enabled once the affected hints are updated.

- - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - -

Close #1329