New: Update `jsdom` to `v13.0.0`
This version of `jsdom` adds support to `canvas` `v2.x`. One of the
most interesting things is that `canvas` downloads now automatically
binaries from `canvas-prebuilt` if they are available or fallbacks
to building and thus solving some of the issues when `canvas-prebuilt`
is not available in the user's platform.

- - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - -

Ref #1445
Fix #1445