Fix: Hide `configuration-all` from the user
This configuration is only for internal consumption to get easy access
to all the documentation and should never be used by a user as it can
cause problems when extending from it.

This PR filters it when running `npm init hintrc` and verifies the
`.hintrc` doesn't `extends` from it.

- - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - -

Fix #3097