Fix: Stability issues with VS Code extension
The VS Code extension could end up breaking `node_modules` in the
workspace or the global shared folder. This was caused because we were
installing `hint@latest` in the workspace and because that process
took too long and the user might closed the editor before it was
ended.
To avoid that we now let the user manage their own `hint` installation
in the workspace, no more prompting about installation. If no version
can be loaded from the workspace, the shared one (`hint@latest`) will
be used.
Also, the shared version is updated after 2 minutes instead of
immediately and only if the shared version has been successfully
loaded.
If the user already has a workspace version there is no point in
updating the shared one as it will not be used.