Fix: Add Node enums required by axe-core
Fixes internal axe-core errors in isVisible when incorrectly
calling getAttribute on a document node. Similarly fixes
axe-core errors calling match on the return value of
getProperty which needed to be the empty string when
no match exists. Both of these issues were causing results
to be omitted for the label rule in certain circumstances.

Also enhances the test helpers to allow passing a list of
tests to validate for each axe-rule (instead of just one pass/fail).
-----------------------------------
Close #5117