Fix: Avoid false positives when using {...spread} in JSX
Introduces a placeholder for `{...spread}` in the transformed HTML along
with a `utils-dom` helper to identify whether additional, unknown attributes
may be present. Uses this helper in `hint-button-type` to avoid triggering
a false-positive.

Also updates JSX-to-HTML conversion to assume a `title` attribute when
`{...spread}` is present if not already set. This ensures a label is provided
for any transformed element with spread, avoiding false-positives in
`hint-axe`.

- - - - - - - - - - - - - - - - - - - -

Fix #5166
Fix microsoft/vscode-edge-devtools/issues/1031
Close #5175